requirements.txt
Django
requests~=2.32.4
Hotel~=1.1
python-decouple~=3.8
dj_database_url>=2.1.0
psycopg2-binary>=2.9.7
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.9.0